+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/block/DisplayPedestal.kt
@@ -31,7 +31,7 @@ class DisplayPedestal : BasePedestal<DisplayPedestalBlockEntity>(BlockUtil.defau
             blockEntity.getPeripheral(Direction.EAST)?.queueEvent(
                 "pedestal_right_click",
                 LuaRepresentation.forItemStack(itemInHand, RepresentationMode.FULL),
-                LuaRepresentation.forItemStack(blockEntity.storedStack, RepresentationMode.FULL)
+                LuaRepresentation.forItemStack(blockEntity.storedStack, RepresentationMode.FULL),
             )
             return InteractionResult.SUCCESS
         }
@@ -48,7 +48,7 @@ class DisplayPedestal : BasePedestal<DisplayPedestalBlockEntity>(BlockUtil.defau
                 ?.queueEvent(
                     "pedestal_left_click",
                     LuaRepresentation.forItemStack(itemInHand, RepresentationMode.FULL),
-                    LuaRepresentation.forItemStack(blockEntity.storedStack, RepresentationMode.FULL)
+                    LuaRepresentation.forItemStack(blockEntity.storedStack, RepresentationMode.FULL),
                 )
         }
         @Suppress("DEPRECATION")