+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/PeripheralProxyPeripheral.kt
@@ -3,10 +3,10 @@ package site.siredvin.peripheralworks.computercraft.peripherals
 import site.siredvin.peripheralium.computercraft.peripheral.owner.BlockEntityPeripheralOwner
 import site.siredvin.peripheralworks.common.blockentity.PeripheralProxyBlockEntity
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
-import site.siredvin.peripheralworks.computercraft.modem.ModemPeripheral
+import site.siredvin.peripheralworks.computercraft.modem.PeripheralHubPeripheral
 
 class PeripheralProxyPeripheral(private val blockEntity: PeripheralProxyBlockEntity) :
-    ModemPeripheral<BlockEntityPeripheralOwner<PeripheralProxyBlockEntity>>(TYPE, BlockEntityPeripheralOwner(blockEntity)) {
+    PeripheralHubPeripheral<BlockEntityPeripheralOwner<PeripheralProxyBlockEntity>>(TYPE, BlockEntityPeripheralOwner(blockEntity)) {
 
     companion object {
         const val TYPE = "peripheral_proxy"